// src/app/models/fleet-type-bu.ts
export class FleetTypeBusinessUnit {
    fleetId: number;
    plantId: number;
    plantCode: string;
    fleetTypeId: number;
    fleetTypeCode: string;
    fleetTypeDesc: string;
    productTypeGroupId: number;
    productTypeGroupCode: string;
    productTypeGroupDesc: string;
    fleetCoverageId: number;
    fleetCoverageCode: string;
    fleetCoverageDesc: string;
    productId: number;
    productCode: string;
    productDesc: string;
    crane: boolean;
    moffett: boolean;
    initialWeight: number;
    finalWeight: number;
    changeDateTime: string;
    userAccountChange: string;
    creationDateTime: string;
    userAccountCreation: string;
}