// src/app/models/fleet-type.ts
export class FleetType {
    fleetTypeId: number;
    fleetTypeCode: string;
    fleetTypeDesc: string;
    comments: string;
    languageCode: string;
    countryCode: string;
    userAccountCreation: string;
    userAccountChange: string;
    creationDateTime: string;
    changeDateTime: string;
}